requirements.txt
psycopg2==2.8.6
kafka-python==2.0.1
requests==2.24.0
